Removed local plantuml.jar file (#1467)
* Find a local plantuml installation

Search the usual location for the jar file and use it instead.
It will work for debian/ubuntu, fedora and arch.

* Nicer layout for the dev-env readme

* Optional tools to make the docs

What's needed to generate the documentation.
How to run it.

* plantuml.jar removed

No need to keep a local copy of the jar-file.
doxygen.sh will find a version, if installed from the package manager.

* Added homebrew support for plantuml